Implement for ConPty
