+++ FluentTerminal.SystemTray/Services/WinPty/WinPtySession.cs
@@ -52,7 +52,7 @@ namespace FluentTerminal.SystemTray.Services.WinPty
                     args = $"\"{request.Profile.Location}\" {args}";
                 }
 
-                spawnConfigHandle = winpty_spawn_config_new(WINPTY_SPAWN_FLAG_AUTO_SHUTDOWN, request.Profile.Location, args, cwd, GetDefaultEnvironmentVariables(), out errorHandle);
+                spawnConfigHandle = winpty_spawn_config_new(WINPTY_SPAWN_FLAG_AUTO_SHUTDOWN, request.Profile.Location, args, cwd, terminalsManager.GetDefaultEnvironmentVariableString(), out errorHandle);
                 if (errorHandle != IntPtr.Zero)
                 {
                     throw new Exception(winpty_error_msg(errorHandle));
@@ -101,23 +101,7 @@ namespace FluentTerminal.SystemTray.Services.WinPty
             Dispose(false);
         }
 
-        private string GetDefaultEnvironmentVariables()
-        {
-            var environmentVariables = Environment.GetEnvironmentVariables();
-            environmentVariables.Add("TERM", "xterm");
-            environmentVariables.Add("TERM_PROGRAM", "FluentTerminal");
-            environmentVariables.Add("TERM_PROGRAM_VERSION", $"{Package.Current.Id.Version.Major}.{Package.Current.Id.Version.Minor}.{Package.Current.Id.Version.Build}.{Package.Current.Id.Version.Revision}");
-
-            var builder = new StringBuilder();
-
-            foreach (DictionaryEntry item in environmentVariables)
-            {
-                builder.Append(item.Key).Append("=").Append(item.Value).Append("\0");
-            }
-            builder.Append('\0');
-
-            return builder.ToString();
-        }
+        
 
         private string GetWorkingDirectory(ShellProfile configuration)
         {
